refactor : 각 기물이 기물 종류를 저장하도록 변경
