+++ src/main/java/chess/domain/board/Board.java
@@ -3,15 +3,15 @@ package chess.domain.board;
 import chess.domain.location.Column;
 import chess.domain.location.Location;
 import chess.domain.location.Row;
-import chess.domain.piece.Bishop;
-import chess.domain.piece.BlackPawn;
+import chess.domain.piece.implement.Bishop;
+import chess.domain.piece.implement.BlackPawn;
 import chess.domain.piece.Color;
-import chess.domain.piece.King;
-import chess.domain.piece.Knight;
+import chess.domain.piece.implement.King;
+import chess.domain.piece.implement.Knight;
 import chess.domain.piece.Piece;
-import chess.domain.piece.Queen;
-import chess.domain.piece.Rook;
-import chess.domain.piece.WhitePawn;
+import chess.domain.piece.implement.Queen;
+import chess.domain.piece.implement.Rook;
+import chess.domain.piece.implement.WhitePawn;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.HashMap;
