+++ src/main/java/chess/domain/piece/PieceType.java
@@ -1,0 +1,5 @@
+package chess.domain.piece;
+
+public enum PieceType {
+    KING, QUEEN, ROOK, KNIGHT, BISHOP, PAWN
+}
