+++ src/main/java/chess/domain/piece/implement/BlackPawn.java
@@ -1,6 +1,7 @@
-package chess.domain.piece;
+package chess.domain.piece.implement;
 
 import chess.domain.board.Path;
+import chess.domain.piece.Color;
 
 public class BlackPawn extends Pawn {
     public BlackPawn() {
