+++ src/main/java/chess/domain/piece/implement/King.java
@@ -1,13 +1,16 @@
-package chess.domain.piece;
+package chess.domain.piece.implement;
 
 import chess.domain.board.Path;
+import chess.domain.piece.Color;
+import chess.domain.piece.Piece;
+import chess.domain.piece.PieceType;
 
 public class King extends Piece {
 
     private static final int MAX_MOVE_DISTANCE = 1;
 
     public King(Color color) {
-        super(color);
+        super(color, PieceType.KING);
     }
 
     @Override
