+++ src/main/java/chess/domain/piece/implement/Rook.java
@@ -1,0 +1,27 @@
+package chess.domain.piece.implement;
+
+import chess.domain.board.Path;
+import chess.domain.piece.Color;
+import chess.domain.piece.Piece;
+import chess.domain.piece.PieceType;
+
+public class Rook extends Piece {
+    public Rook(Color color) {
+        super(color, PieceType.ROOK);
+    }
+
+    @Override
+    public boolean canMove(Path path) {
+        if (path.hasPiecePathExcludedTarget()) {
+            return false;
+        }
+        if (path.containsDiagonalDirection()) {
+            return false;
+        }
+        return path.isNotAllyAtTarget() && path.hasCountOfDirection(1);
+    }
+
+    @Override
+    public void move() {
+    }
+}
